+++ src/plang/translator/codegen/CodeData.java
@@ -2,9 +2,9 @@ package plang.translator.codegen;
 
 public final class CodeData {
     public final byte[] code;
-    public final int[] constantPool;
+    public final long[] constantPool;
 
-    public CodeData(byte[] code, int[] constantPool) {
+    public CodeData(byte[] code, long[] constantPool) {
         this.code = code;
         this.constantPool = constantPool;
     }
